Experimentation/3d-point-cloud-weave/pyproject.toml
[project]
name = "3d-point-cloud-weave"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "moderngl>=5.12.0",
    "numpy>=2.4.1",
    "pillow>=12.1.0",
    "pyside6>=6.10.1",
    "torch>=2.10.0",
]
